test pre-commit script for enforcing code standards
to try it out,
  ln -s ../../scripts/pre-commit.sh .git/hooks/pre-commit
make sure that it is executable (chmod u+x scripts/pre-commit.sh) and
install astyle. On windows, it suffices to copy AStyle.exe to
C:\Program Files\git\bin
Why astyle? Because, unlike indent or clang-format, it is possible to
change /just/ whitespace, and not necessarily all the other things,
like brackets, padding etc.